# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: "Inthub Backend API"
  version: "1.0.0"
  description: >
    この仕様書は、FastAPI によるバックエンド実装に対応する API を定義しています。
    各エンドポイントには、詳細な説明、想定されるエラー(400, 401, 404, 409, 500 など)も記載しています。
servers:
  - url: "https://inthub-fjfjehdsc4akamdg.japaneast-01.azurewebsites.net"
...
paths:
...
  /api/v1/company:
    post:
      tags:
        - Company
      summary: "企業登録"
      description: >
        新規企業を登録します。  
        リクエストボディには企業名とサーバー名を含みます。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCompany"
      responses:
        '200':
          description: "企業登録に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyRegisterResponse"
        '400':
          description: "入力データが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: "既に存在する企業"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Company
      summary: "全テナント情報取得"
      description: "全てのテナント（企業）情報を取得します。"
      responses:
        '200':
          description: "テナント情報の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/company/{company_id}:
    get:
      tags:
        - Company
      summary: "指定企業のテナント情報取得"
      description: >
        指定した company_id に対応するテナント情報を取得します。
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
          description: "企業の一意識別子"
      responses:
        '200':
          description: "テナント情報の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInfoResponse"
        '404':
          description: "指定した企業が見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/company/update-schemas:
    post:
      tags:
        - Company
      summary: "全テナントスキーマ更新"
      description: >
        全てのテナントデータベースのスキーマを更新します。  
        実行には時間がかかる可能性があるため注意が必要です。
      responses:
        '200':
          description: "スキーマ更新に成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/company/{company_id}/senpai/{senpai_id}:
    get:
      tags:
        - Company
      summary: "先輩詳細取得"
      description: >
        指定した企業と先輩IDに基づいて先輩の詳細情報を取得します。
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
          description: "企業の一意識別子"
        - in: path
          name: senpai_id
          required: true
          schema:
            type: string
          description: "先輩の一意識別子"
      responses:
        '200':
          description: "先輩詳細の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSenpaiDetailResponse"
        '404':
          description: "先輩情報が見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/company/{company_id}/senpai:
    get:
      tags:
        - Company
      summary: "先輩一覧取得"
      description: >
        指定した企業に紐づく全先輩情報の一覧を取得します。
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
          description: "企業の一意識別子"
      responses:
        '200':
          description: "先輩一覧の取得に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SenpaiListResponse"
        '404':
          description: "企業または先輩情報が見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/company/{company_id}/data/upload:
    post:
      tags:
        - Company
      summary: "データアップロード"
      description: >
        会社IDを指定して、データをアップロードします。
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
          description: "企業の一意識別子"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: "データアップロードに成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDataResponse"
        '400':
          description: "入力データが不正"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "指定した企業が見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/company/{company_id}/data/delete:
    delete:
      tags:
        - Company
      summary: "データ削除"
      description: >
        会社IDを指定して、データを削除します。
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: string
          description: "企業の一意識別子"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDataRequest"
      responses:
        '200':
          description: "データ削除に成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDataResponse"
        '400':
          description: "無効なリクエスト"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "指定した企業またはデータが見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "サーバ内部エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterCompany:
      type: object
      properties:
        company_name:
          type: string
        company_server_name:
          type: string
      required:
        - company_name
        - company_server_name
    CompanyRegisterResponse:
      type: object
      properties:
        company_id:
          type: string
        status:
          type: string
        message:
          type: string
      required:
        - company_id
        - status
        - message
...
    TenantInfoFromTenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        company_server_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - company_name
        - company_server_name
        - created_at
        - updated_at
    TenantInfoResponse:
      type: object
      properties:
        status:
          type: string
        tenant:
          $ref: "#/components/schemas/TenantInfoFromTenant"
      required:
        - status
        - tenant
    TenantInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        company_id:
          type: string
          format: uuid
        company_server_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - company_name
        - company_id
        - company_server_name
        - created_at
        - updated_at
    TenantListResponse:
      type: object
      properties:
        status:
          type: string
        tenants:
          type: array
          items:
            $ref: "#/components/schemas/TenantInfo"
      required:
        - status
        - tenants
...
    GetSenpaiDetailResponse:
      type: object
      properties:
        senpai_id:
          type: string
        senpai_name:
          type: string
        profile:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required:
        - senpai_id
        - senpai_name
        - profile
        - created_at
    SenpaiListResponse:
      type: object
      properties:
        status:
          type: string
        senpais:
          type: array
          items:
            $ref: "#/components/schemas/GetSenpaiDetailResponse"
      required:
        - status
        - senpais
    UploadDataResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        fileId:
          type: string
      required:
        - status
        - message
        - fileId
...
    DeleteDataResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status
        - message
    DeleteDataRequest:
      type: object
      properties:
        file_id:
          type: string
      required:
        - file_id
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        error_message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - error_code
        - error_message